\name{c060-internal}
\title{Internal c060 functions}
\alias{my.balanced.folds}

\description{Internal c060 functions}
\usage{
my.balanced.folds(class.column.factor, cross.outer)

}
%\author{}
\details{These are not intended for use by users. \code{my.balanced.folds}
 produces balanced stratified folds, which are used for tuning penalization parameters.
}
\keyword{internal}
