% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c060-package.R
\docType{package}
\name{c060-package}
\alias{c060-package}
\alias{c060}
\title{c060: Extended Inference for Lasso and Elastic-Net Regularized Cox and Generalized Linear Models}
\description{
The c060 package provides additional functions to perform stability selection, model validation and parameter tuning for glmnet models.
}
\references{
Friedman, J., Hastie, T. and Tibshirani, R. (2008)
\emph{Regularization Paths for Generalized Linear Models via Coordinate
Descent}, \url{https://web.stanford.edu/~hastie/Papers/glmnet.pdf}\cr
\emph{Journal of Statistical Software, Vol. 33(1), 1-22 Feb 2010}\cr
\url{https://www.jstatsoft.org/v33/i01/}\cr Simon, N., Friedman, J., Hastie,
T., Tibshirani, R. (2011) \emph{Regularization Paths for Cox's Proportional
Hazards Model via Coordinate Descent, Journal of Statistical Software, Vol.
39(5) 1-13}\cr \url{https://www.jstatsoft.org/v39/i05/}\cr Porzelius, C.,
Binder, H., and Schumacher, M. (2009) \emph{Parallelized prediction error
estimation for evaluation of high-dimensional models, Bioinformatics, Vol.
25(6), 827-829.}\cr Sill M., Hielscher T., Becker N. and Zucknick M. (2014),
\emph{c060: Extended Inference with Lasso and Elastic-Net Regularized Cox
and Generalized Linear Models, Journal of Statistical Software, Volume
62(5), pages 1--22.} https://doi.org/10.18637/jss.v062.i05.
}
\seealso{
\code{\link[c060]{predictProb.coxnet}},
\code{\link[peperr]{peperr}}, \code{\link[glmnet]{glmnet}}
}
\author{
Thomas Hielscher \email{t.hielscher@dkfz.de}, Manuela Zucknick, Natalia Becker, Frédéric Bertrand.
}
\keyword{internal}
\keyword{models}
\keyword{penalized}
\keyword{regression}
\keyword{survival}
