% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_intsearch.R
\name{coef.sum.intsearch}
\alias{coef.sum.intsearch}
\title{Get coefficients for a model}
\usage{
\method{coef}{sum.intsearch}(object, ...)
}
\arguments{
\item{object}{an object as returned by the function \code{summary.intsearch}.}

\item{\dots}{additional argument(s)}
}
\value{
named vector of non-zero coeficients for the optimal lambda
}
\description{
Get coefficients for a model after applying interval search for tuning
parameters
}
\references{
Sill M., Hielscher T., Becker N. and Zucknick M. (2014),
\emph{c060: Extended Inference with Lasso and Elastic-Net Regularized Cox
and Generalized Linear Models, Journal of Statistical Software, Volume
62(5), pages 1--22.} https://doi.org/10.18637/jss.v062.i05.
}
\seealso{
\code{\link[c060]{EPSGO}},
\code{\link[c060]{summary.intsearch}},\code{\link[c060]{plot.sum.intsearch}}
}
\author{
Natalia Becker \ \email{natalia.becker@dkfz.de}
}
\keyword{system}
